Simplified filesystem streams (#993)
* Simplified filesystem streams

* Updated UPGRADE.md

* Updated dependencies

* Fixed failing examples

+++ src/core/etl/src/Flow/ETL/Filesystem/LocalFilesystem.php
@@ -66,6 +66,25 @@ final class LocalFilesystem implements Filesystem
         return false;
     }
 
+    public function mv(Path $from, Path $to) : void
+    {
+        if (!$from->isLocal() || !$to->isLocal()) {
+            throw new RuntimeException(\sprintf('Paths "%s" and "%s" are not local', $from->uri(), $to->uri()));
+        }
+
+        if ($from->isPattern() || $to->isPattern()) {
+            throw new RuntimeException('Pattern paths can\'t be moved');
+        }
+
+        if (\file_exists($to->path())) {
+            $this->rm($to);
+        }
+
+        if (!\rename($from->path(), $to->path())) {
+            throw new RuntimeException(\sprintf('Can\'t move "%s" to "%s"', $from->uri(), $to->uri()));
+        }
+    }
+
     public function open(Path $path, Mode $mode) : FileStream
     {
         if (!$path->isLocal()) {
@@ -95,7 +114,9 @@ final class LocalFilesystem implements Filesystem
             if (\is_dir($path->path())) {
                 $this->rmdir($path->path());
             } else {
-                \unlink($path->path());
+                if (\file_exists($path->path())) {
+                    \unlink($path->path());
+                }
             }
 
             return;
